syntax: fix erroneous generic type highlighting
Don't allow angle brackets (`<...>`) to match outside of expected
locations (composite types, function calls) due to how this interacts
with the comparison operators `<` and `>`. The vim matcher will
consider any pair of `<` and `>` operators to be eligible, in many
places that it should not. For example:

    defn foo<?T> (t: ?T) -> True|False:
        if SOME-GLOBAL < 10:
            true
        else if SOME-GLOBAL > 20:
            false

In the above example, vim will correctly mark `<?T>` as a pair of
generic type brackets -- however, it will also incorrectly mark from the
less-than operator in `if SOME-GLOBAL < 10` all the way to the
greater-than operator in `else if SOME-GLOBAL > 20:`.

Hence, we must only match this syntax region in positions where we know
it can occur.