syntax: don't highlight multi-line composite types
Causes too many highlighting errors to be worth it, due to how the
syntax engine works. Unfortunately, there is a tradeoff between being
able to highlight multi-line composite types where the type operator
appears on the following line, e.g:

   Long-And-Complicated-Type-Here
    | Return-Type

and being able to properly highlight indent-sensitive syntax groups --
regions that require the syntax engine to match on the leading indent,
e.g:

    match(foo):
      (b: Bar)
    ^^ ---------- checked for, to avoid false positives

The specific issue that arises occurs when a type appears as the last
syntax element on a line preceding a line containing an indent-sensitive
syntax group. In such a scenario, the indent-sensitive syntax group is
not properly highlighted, e.g:

    match(foo):
      (a: A): v is False
      (b: B):
          ^ ------------- should be highlighted, but isn't

This is due to how composite type highlighting works, specifically
because of its use of `nextgroup` and `skip*`. When the syntax engine
sees that we've matched a type, it attempts to match the groups
specified in `nextgroup` at the position following the type (in our
case, to check for a type operator). If the groups fail to match, and
`skipwhite` is specified, it will first skip any following whitespace,
and if `skipnl` is specified, it will also skip new lines.

This is an issue, because when this behavior is triggered before an
indent-sensitive syntax group, it causes the syntax engine to skip over
the indent that is necessary to properly highlight that syntax group,
causing the highlighting to not work correctly.

Ultimately the indent-sensitive syntax groups seem to be more prevalent
and useful (`match` arms) than being able to support multi-line
composite types that have the type operator on the next line.