syntax: fix `match` highlighting with single-clause form
Previous changes attempted to fix bugs with `stanzaMatchClause`, where it would match on any pair of parentheses within a `match` expression, e.g:

   match(foo):
     (bar: B):
        a(bar)
          ^^^----- was matched as `stanzaMatchClause`,

Which, besides being incorrect, also caused a variety of bizarre
matching/highlighting issues inside function calls.

These issues were fixed by requiring preceding whitespace before the
opening parenthesis of the `match` clause. This breaks the following
case:

   match(foo: String):
        ^-------------- no preceding whitespace here!

So, instead of using the preceding whitespace test, we instead invent
the following hack: any pair of parentheses that contains a colon (`:`)
is a `stanzaMatchClause`, e.g:

  match(foo: String):
        ^^^^^^^^^^^^--- highlighted as a `match` clause
           ^----------- due to this colon

  match(foo):
    (foo: String):
    ^^^^^^^^^^^^^------ highlighted as a `match` clause
        ^-------------- due to this colon
    (bar):
       foo()
          ^^----------- not `match` clause, no colon

This is *technically* wrong from a syntax point of view, because
`(bar):` in the preceding snippet is technically a `match` clause, but
this causes no issues for highlighting, so we can ignore it.